+++ pyCC/pyCmp/asmgen.py
@@ -1,4 +1,4 @@
-from parser import *
+from .parser import *
 
 def asmgenerate(ast: ProgramNode):
     asmTree = ast.assemble()
