+++ pyCC/pyCmp/lexer.py
